add security with tokens and protect views

+++ controllers/TaskController.js
@@ -1,27 +1,37 @@
-const User = require("../models/User");
+const Users = require("../models/User");
 const Tasks = require("../models/Tasks");
 const { Op } = require("sequelize");
 
 //Home controllers
 exports.GetHome = async (req, res, next) => {
-  const arrayTasks = await Tasks.findAll({ where: { userId: 1 }, order: [["createdAt", "DESC"]] });
-  const tasks = arrayTasks.map((result) => result.dataValues);
+  let tasks;
+  let user;
+
+  if (req.user) {
+    const arrayTasks = await Tasks.findAll({
+      where: { userId: req.user.id },
+      order: [["createdAt", "DESC"]],
+    });
+    tasks = arrayTasks.map((result) => result.dataValues);
+
+    user = await Users.findOne({where: {id: req.user.id }});
+  }
 
   res.render("client/home", {
     pageTitle: "Home",
     homeActive: true,
-    place: 'home',
+    place: "home",
     tasks,
+    user,
   });
 };
 
 exports.PostCreateTask = (req, res, next) => {
   const content = req.body.Task;
-  const userId = req.body.UserId;
 
   Tasks.create({
     content: content,
-    userId: userId,
+    userId: req.user.id,
   })
     .then(() => {
       res.redirect("/home");
@@ -33,9 +43,9 @@ exports.PostCreateTask = (req, res, next) => {
 
 exports.PostCompletedTask = async (req, res, next) => {
   let taskId = req.query.taskId;
-  let place = req.params.Place
+  let place = req.params.Place;
 
-  const task = await Tasks.findOne({ where: { id: taskId } });
+  const task = await Tasks.findOne({ where: { id: taskId, userId: req.user.id } });
 
   const isCompleted = task.dataValues.isCompleted;
 
@@ -60,9 +70,9 @@ exports.PostCompletedTask = async (req, res, next) => {
 
 exports.PostDeleteTask = async (req, res, next) => {
   const taskId = req.body.TaskId;
-  let place = req.params.Place
+  let place = req.params.Place;
 
-  Tasks.destroy({where: {id: taskId}})
+  Tasks.destroy({ where: { id: taskId, userId: req.user.id } })
     .then(() => {
       res.redirect(`/${place}`);
     })
@@ -74,9 +84,9 @@ exports.PostDeleteTask = async (req, res, next) => {
 exports.PostEditTask = (req, res, next) => {
   const content = req.body.Task;
   const taskId = req.body.TaskId;
-  let place = req.params.Place
+  let place = req.params.Place;
 
-  Tasks.update({content: content}, {where: {id: taskId}})
+  Tasks.update({ content: content }, { where: { id: taskId, userId: req.user.id } })
     .then(() => {
       res.redirect(`/${place}`);
     })
@@ -87,9 +97,10 @@ exports.PostEditTask = (req, res, next) => {
 
 //Completed controllers
 exports.GetCompletedTasks = async (req, res, next) => {
-  const arrayTasks = await Tasks.findAll({ 
-    where: { userId: 1, [Op.and]: [{isCompleted: true}], }, 
-    order: [["createdAt", "DESC"]] });
+  const arrayTasks = await Tasks.findAll({
+    where: { userId: req.user.id, [Op.and]: [{ isCompleted: true }] },
+    order: [["createdAt", "DESC"]],
+  });
   const tasks = arrayTasks.map((result) => result.dataValues);
 
   res.render("client/completed&pending", {
@@ -97,21 +108,22 @@ exports.GetCompletedTasks = async (req, res, next) => {
     completedActive: true,
     tasks,
     completedTask: true,
-    place: 'completed-tasks',
+    place: "completed-tasks",
   });
 };
 
 //Pending controllers
 exports.GetPendingTasks = async (req, res, next) => {
-  const arrayTasks = await Tasks.findAll({ 
-    where: { userId: 1, [Op.and]: [{isCompleted: {[Op.not]: true}}], }, 
-    order: [["createdAt", "DESC"]] });
+  const arrayTasks = await Tasks.findAll({
+    where: { userId: req.user.id, [Op.and]: [{ isCompleted: { [Op.not]: true } }] },
+    order: [["createdAt", "DESC"]],
+  });
   const tasks = arrayTasks.map((result) => result.dataValues);
 
   res.render("client/completed&pending", {
     pageTitle: "Pending tasks",
     pendingActive: true,
     tasks,
-    place: 'pending-tasks'
+    place: "pending-tasks",
   });
 };
